+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Library pedestal.immutant has been removed
 - Library pedestal.tomcat has been removed
 - Namespace io.pedestal.http.jetty.websockets has been removed, and replaced with io.pedestal.http.websockets
+- Namespace io.pedestal.interceptor.error has been moved to a new library, io.pedestal/pedestal.error
+- Library pedestal.service no longer has a dependency on org.clojure/core.match; the pedestal.error library has that dependency
 
 Other changes: